requirements.txt
flask==2.3.2
gunicorn==21.2.0
opencv-python==4.9.0.80
numpy==1.24.4
Pillow==10.1.0
streamlit==1.32.2
requests==2.31.0